backend/requirements.txt
asgiref==3.5.2
autopep8==2.0.0
backports.zoneinfo==0.2.1;python_version<"3.9"
Django==4.1.2
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
importlib-metadata==5.0.0
Markdown==3.4.1
Pillow==9.3.0
pycodestyle==2.9.1
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1
zipp==3.10.0
